requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain-chroma
langchain-text-splitters
sentence-transformers
python-dotenv
ipykernel
pdfplumber
pathlib
tabulate
langchain-groq